edr/TA/requirements.txt
pyzmq
protobuf
pytest-html
datetime
sseclient
psutil
